Consolidate imports
- Make sure `client` is always imported from "@/api"
- Use inline type import
