+++ client/src/api/schema/index.ts
@@ -1,3 +1,2 @@
-export { fetcher } from "./fetcher";
-export { client, GalaxyApiPaths } from "./fetcher";
-export type { components, operations, paths } from "./schema";
+export { client, type GalaxyApiPaths } from "./fetcher";
+export { type components } from "./schema";
