+++ client/src/api/tags.ts
@@ -1,4 +1,4 @@
-import { client, type components } from "@/api/schema";
+import { client, type components } from "@/api";
 import { rethrowSimple } from "@/utils/simple-error";
 
 type TaggableItemClass = components["schemas"]["TaggableItemClass"];
