+++ client/src/components/ObjectStore/Templates/types.ts
@@ -1,4 +1,4 @@
-import { type components } from "@/api/schema/schema";
+import { type components } from "@/api/schema";
 
 export type ObjectStoreTemplateSummary = components["schemas"]["ObjectStoreTemplateSummary"];
 export type ObjectStoreTemplateSummaries = ObjectStoreTemplateSummary[];
